GEODE-5490: Split up TxCommitMessageBackwardsCompatibilityDUnitTest
